+++ lib/dmtd_vbmapp_data/protocol_area.rb
@@ -2,8 +2,8 @@ require 'dmtd_vbmapp_data/request_helpers'
 
 module DmtdVbmappData
 
-  # Provides for the retrieving of VB-MAPP area information from the VB-MAPP Data Server.
-  class VbmappArea
+  # Provides for the retrieving of VB-MAPP Protocol Area information from the VB-MAPP Data Server.
+  class ProtocolArea
 
     # @!attribute [r] client
     #   @return [Client] the associated client
@@ -19,7 +19,7 @@ module DmtdVbmappData
     #
     # @option opts [Client] :client A client instance
     # @option opts [Hash] :area_index_json The vbmapp area index json for the VB-Mapp Area in the format described at
-    #     {https://datamtd.atlassian.net/wiki/pages/viewpage.action?pageId=18710543 /1/vbmapp/index - GET REST api - Area Fields}
+    #     {https://datamtd.atlassian.net/wiki/pages/viewpage.action?pageId=18710543 /1/protocol/index - GET REST api - Area Fields}
     def initialize(opts)
       @client = opts.fetch(:client)
 
@@ -30,10 +30,10 @@ module DmtdVbmappData
 
     # @note This method does *not* block on server access
     #
-    # @return [Array<VbmappAreaGroup>] all of the VB-Mapp's {VbmappAreaGroup} instances
+    # @return [Array<ProtocolAreaGroup>] all of the VB-Mapp's {ProtocolAreaGroup} instances
     def groups
       @groups = @groups_index.map.with_index do |group_index_json, group_num|
-        VbmappAreaGroup.new(client: client, area: area, group_index_json: group_index_json)
+        ProtocolAreaGroup.new(client: client, area: area, group_index_json: group_index_json)
       end if @groups.nil?
 
       @groups