+++ spec/protocol_area_spec.rb
@@ -2,10 +2,10 @@ require 'spec_helper'
 
 module DmtdVbmappData
 
-  describe VbmappArea do
+  describe ProtocolArea do
 
     it 'can be created' do
-      area = VbmappArea.new(client: client, area_index_json: area_index_json)
+      area = ProtocolArea.new(client: client, area_index_json: area_index_json)
       expect(area).to_not be nil
       expect(area.area).to eq(:milestones)
     end
@@ -18,14 +18,14 @@ module DmtdVbmappData
       expect(areas.size).to be > 0
 
       areas.each do |area|
-        expect(area.is_a?(DmtdVbmappData::VbmappArea)).to eq(true)
+        expect(area.is_a?(DmtdVbmappData::ProtocolArea)).to eq(true)
       end
     end
 
     private
 
     def client
-      Client.new(id: 57)
+      Client.new(id: VBMDATA_TEST_CLIENT_ID)
     end
 
     def area_index_json